+++ packages/ember-data/lib/system/request.js
@@ -2,7 +2,7 @@
   @module ember-data
 */
 
-var get = Ember.get, set = Ember.set;
+var get = Ember.get;
 
 /**
   @class Request
@@ -13,12 +13,13 @@ DS.Request = Ember.Object.extend({
 
   store: null,
   type: null,
-  promise: null,
-  resolve: null,
-  reject: null,
-  sinceToken: null,
+  query: null,
+  fetchPage: null,
   page: null,
   pageSize: null,
+  sinceToken: null,
+
+  deferred: null,
 
   init: function() {
     this.deferred = Ember.RSVP.defer();
@@ -29,7 +30,7 @@ DS.Request = Ember.Object.extend({
     var store = get(this, 'store'),
         type = get(this, 'type'),
         query = get(this, 'query');
-    Ember.assert('You tried to call Request.loadPage but no fetchPage method has been provided', this.fetchPage);
+    Ember.assert('You tried to call Request.loadPage but no fetchPage function has been provided', this.fetchPage);
     return this.fetchPage(store, type, query, page);
   }
